+++ react-todo-list/src/components/NewToDo/NewToDo.css
@@ -1,0 +1,6 @@
+.NewToDo {
+    width: 50%;
+    margin: 16px auto;
+    padding: 16px;
+    text-align: center;
+}
